backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
requests==2.31.0
Werkzeug==2.3.7
python-dotenv==1.0.0